// Part 5 - Transitions/src/interfaces/filter.interface.ts
type Icon = 'bi-binoculars-fill' | 'bi-tags-fill' | 'bi-coin';
type InputType = 'text' | 'checkbox' | 'radio';
export type CategoryName = 'Gadgets Lumineux' | 'Technologie Futuriste' | 'Café et Code';
type PriceName = 'Tous les prix' | 'Moins de 100€' | 'Entre 100 et 200€' | 'Plus de 200€';
type PriceType = 'all' | '0-100' | '100-200' | '200+';

interface FilterInterface {
    title: string;
    icon: Icon;
    inputType: InputType;
    placeholder?: string;
    value: string | CategoryName[];
    items?: {
        id: number;
        name: string;
        value?: string;
    }[];
}

interface SearchFilterInterface extends FilterInterface {
    inputType: 'text';
    value: string;
    placeholder: 'Chercher';
}


interface CategoriesFilterInterface extends FilterInterface {
    inputType: 'checkbox';
    value: CategoryName[];
    items: {
        id: number;
        name: CategoryName;
    }[];
}



interface PricesFilterInterface extends FilterInterface {
    inputType: 'radio';
    value: PriceType;
    items: {
        id: number;
        name: PriceName;
        value: PriceType;
    }[];
}

export interface FiltersInterface {
    search: SearchFilterInterface;
    categories: CategoriesFilterInterface;
    prices: PricesFilterInterface;
}